Error 'undefined formAutocomplete' displayed immediately when selecting Common: Navigate/ShowDialog action #2696
