+++ shesha-reactjs/src/components/configurableItemAutocomplete/formAutocomplete.tsx
@@ -1,0 +1,22 @@
+import { FC } from 'react';
+import { GenericConfigItemAutocomplete, ConfigurableItemAutocompleteRuntimeProps, StandardAutocompleteProps } from './generic';
+import React from 'react';
+import { FormFullName } from '@/interfaces';
+
+export type IFormAutocompleteRuntimeProps = ConfigurableItemAutocompleteRuntimeProps<FormFullName, Omit<StandardAutocompleteProps, 'entityType' | 'filter'>>;
+
+const FORM_CONFIG_ENTITY_TYPE = 'Shesha.Core.FormConfiguration';
+
+const baseFormFilter = {
+    "==": [{ "var": "isTemplate" }, false]
+};
+
+export const FormAutocomplete: FC<IFormAutocompleteRuntimeProps> = (props) => {
+    return (
+        <GenericConfigItemAutocomplete
+            {...props}
+            entityType={FORM_CONFIG_ENTITY_TYPE}
+            filter={baseFormFilter}
+        />
+    );
+};
