+++ shesha-reactjs/src/components/configurableItemAutocomplete/notificationAutocomplete.tsx
@@ -1,0 +1,17 @@
+import { FC } from 'react';
+import { FormFullName } from '@/providers/form/models';
+import { GenericConfigItemAutocomplete, ConfigurableItemAutocompleteRuntimeProps, StandardAutocompleteProps } from './generic';
+import React from 'react';
+
+export type INotificationAutocompleteRuntimeProps = ConfigurableItemAutocompleteRuntimeProps<FormFullName, Omit<StandardAutocompleteProps, 'entityType' | 'filter'>>;
+
+const NOTIFICATION_CONFIG_ENTITY_TYPE = 'Shesha.Domain.NotificationTypeConfig';
+
+export const NotificationAutocomplete: FC<INotificationAutocompleteRuntimeProps> = (props) => {
+    return (
+        <GenericConfigItemAutocomplete
+            {...props}
+            entityType={NOTIFICATION_CONFIG_ENTITY_TYPE}
+        />
+    );
+};
