+++ shesha-reactjs/src/components/index.ts
@@ -19,8 +19,8 @@ export { default as BooleanDropDown } from './booleanDropDown';
 export { SidebarContainer } from './sidebarContainer';
 export { default as AutoCompletePlaces } from './googlePlacesAutocomplete';
 export { EndpointsAutocomplete } from './endpointsAutocomplete/endpointsAutocomplete';
-export { FormAutocomplete } from './formAutocomplete';
-export { NotificationAutocomplete } from './notificationAutocomplete';
+export { FormAutocomplete } from './configurableItemAutocomplete/formAutocomplete';
+export { NotificationAutocomplete } from './configurableItemAutocomplete/notificationAutocomplete';
 export { default as CustomFile } from './customFile';
 export { default as ErrorBoundary } from './errorBoundary/errorBoundary';
 export { default as CustomErrorBoundary } from './customErrorBoundary';
@@ -118,3 +118,4 @@ export { ListEditor } from './listEditor';
 export { ButtonGroupConfigurator } from './buttonGroupConfigurator';
 export { ConfigurableComponentRenderer } from "./configurableComponentRenderer";
 export { PermissionAutocomplete, type IPermissionAutocompleteProps } from './permissionAutocomplete';
+export { ConfigItemAutocomplete } from './configurableItemAutocomplete';
