+++ shesha-reactjs/src/components/modelConfigurator/viewsEditor/index.tsx
@@ -1,4 +1,4 @@
-import { FormAutocomplete } from '@/components/formAutocomplete';
+import { FormAutocomplete } from '@/components/configurableItemAutocomplete/formAutocomplete';
 import React, { FC, useMemo } from 'react';
 import {
     Button,
