+++ shesha-reactjs/src/designer-components/_legacyComponents/listControl/settingsv2.tsx
@@ -3,7 +3,7 @@ import React, { FC, useState } from 'react';
 import { QueryBuilderRenderer } from '@/designer-components/queryBuilder/queryBuilderRenderer';
 import { QueryBuilderWithModelType } from '@/designer-components/queryBuilder/queryBuilderWithModelType';
 import { Autocomplete } from '@/components/autocomplete';
-import { FormAutocomplete } from '@/components/formAutocomplete';
+import { FormAutocomplete } from '@/components/configurableItemAutocomplete/formAutocomplete';
 import PropertyAutocomplete from '../../../components/propertyAutocomplete/propertyAutocomplete';
 import SectionSeparator from '@/components/sectionSeparator';
 import Show from '@/components/show';
