+++ shesha-reactjs/src/designer-components/configurableActionsConfigurator/genericArgumentsEditor.tsx
@@ -11,6 +11,7 @@ export interface IProps<TModel extends IConfigurableActionArguments> {
   onCancel: () => void;
   onValuesChange?: (changedValues: any, values: TModel) => void;
   readOnly?: boolean;
+  cacheKey?: string;
 }
 
 function GenericArgumentsEditor<TModel extends IConfigurableActionArguments>({
@@ -19,6 +20,7 @@ function GenericArgumentsEditor<TModel extends IConfigurableActionArguments>({
   markup,
   onValuesChange,
   readOnly = false,
+  cacheKey,
 }: IProps<TModel>) {
   const formRef = useShaFormRef();
 
@@ -34,6 +36,7 @@ function GenericArgumentsEditor<TModel extends IConfigurableActionArguments>({
       shaFormRef={formRef}
       onFinish={onSave}
       markup={markup}
+      cacheKey={cacheKey}
       initialValues={model}
       onValuesChange={onValuesChange}
     />
